Start reworking recipes.
Recipes should be split into two categories: Blueprints, like traditional
crafting; and ingredients, covering things like dyes. However, two different
interfaces for recipes is lame, so we're going to move the actual logic for
checking whether a recipe matches into the interface and its implementors.
This should let us move a buttload of code out of inventory, which is becoming
bloated, and also give plugin authors a lot more freedom when writing new
recipes.

This is obstensibly only covered by #336, which is a very misleading bug.