ibravo, serializers: Add ISerializer.connect().
Explicit is better than implicit; things shouldn't be started at
initialization time. This patch removes the world_url Exocet parameter and
replaces it with an explicit .connect() method for serializers.

Ironically, we could have done this in the very beginning. It somehow simply
never fucking occurred to me.

Related to #365.